requirements.txt
automata-lib
numpy
graphviz
sklearn
tqdm